MONGOID-5052 embedded associations can be "eager loaded"
In actuality, eager loading an embedded association does nothing; this is
primarily useful only for chaining eager loading together so that you can
eager load a referenced association via an embedded association.
