+++ lib/mongoid/association/embedded.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'mongoid/association/embedded/cyclic'
+require 'mongoid/association/embedded/eager'
 require 'mongoid/association/embedded/embedded_in'
 require 'mongoid/association/embedded/embeds_many'
 require 'mongoid/association/embedded/embeds_one'
